Separate code translation from parsing
